Changed App structure for Deployment
 - Project root is now node app root
 - Node serves frontend on port 80
 - Node runs server on port 4242
